REFACTOR: 중복코드 제거, 네이밍 변경, 코드 구조 변경
1. BaseInputsValidator 클래스 선언 및 상속 통해서 vm과 mock의 중복 코드 제거
2. mock 클래스 네이밍 변경
3. AuthenticationPresenterTests private 부분 분리
